move `trim_offset` to `string_utils`
the diff viewer will need it in a future commit
